pyproject.toml
[project]
name = "korix-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.7",
    "fastapi>=0.136.0",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.2",
    "langgraph>=1.1.9",
    "psycopg2-binary>=2.9.12",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "scalar-fastapi>=1.8.2",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
